feat: 요구사항4 구현
